Update README with technical notes and basic customization
This adds more technical notes on how to use Git with the project,
and a new section on basic customization that details the currently
used markdown files and any technical notes associated with editing
them. Specifically, to resolve some confusion over how to add a
picture for faculty members a technical note was added that details
where their associated image should be stored in the repository and
how Hugo links to it.

This will be expanded, and a section on Advanced Customization will
be added shortly.